star-bot/requirements.txt
discord.py
sqlite3
dotenv
pytest